issue refs TP-145: migration

+++ Orchard-1.9/src/Orchard.Web/Modules/Teeyoot.Module/Models/PromotionRecord.cs
@@ -12,7 +12,7 @@ namespace Teeyoot.Module.Models
 
         public virtual string DiscountType { get; set; }
 
-        public virtual int AmountSize { get; set; }
+        public virtual double AmountSize { get; set; }
 
         public virtual string AmountType { get; set; }
